Modify the format of the keyword tokens
The tokenizer tokenized keywords like `msgid` or `msgstr` as, e.g., `{:keyword,
line, :msgid}`. However, tagging a keyword as `:keyword` is useless since an
hypothetical parsing grammar (be it a yecc-based grammar or a handwritten
parser) will still need to know what keyword a `:keyword` token holds.

I changed the behaviour of the tokenizer so that it now tokenizes keywords as
`{keyword, line}` (for example, `{:msgid, 3}`).