Add a failing test for Gettext.PO.Parser
I believe this failure is caused by yecc behaving weirdly when `{:string, line,
value}` tokens are used and the value is a binary. I keep getting
`FunctionClauseError`s for `:io_lib.write_string/1` which gets called with the
value of the token, failing because it expects a char list and it gets a binary
instead.