Fix a bug caused by yecc
The commit title is bold, but I think I found the bug highlighted by the failing
test from the previous commit. yecc interprets `{string, Line, Contents}` tokens
differently than it handles other tokens as it can be seen at
https://github.com/erlang/otp/blob/maint/lib/parsetools/src/yecc.erl#L2496.

This caused problems when using binaries (common in Elixir) instead of char
lists since yecc tries to "print" them when a syntax error occurs, and for
printing them it uses `io_lib:write_string/1` which expects a char list and
fails with a binary.

The fix to this problem was to just use `{str, Line, Contents}` tokens in
`Gettext.PO.Tokenize.tokenize/1`.