Compile .po files into function clauses
This commit lays the foundations of .po files compilation. A `__using__` macro
has been added to the `Gettext` module in order to compile .po files and
dynamically generate function clauses (in the caller module) in the form:

    def lgettext("pt_BR", "default", "Hello world") do
      "Ola mundo"
    end

In order to test that everything works correctly, I had to add dummy
applications inside test/fixtures/test_applications. These applications only
have an `ebin/` directory (empty except for a .gitkeep file for keeping them in
git) and translations in them.