Properly raise a File.Error when parsing files
If parsing the file returned an error *with the file*, the solution was to use
`File.read!/1` in order to raise that error. This was extremely unclear and
extremely error prone; for example, what if the error disappeared between the
call to `parse_file/1` and the call to `File.read!/1`?

I fixed that by manually raising a `File.Error` error with the proper metadata.