Add a `Gettext.PO.PluralTranslation` struct
Instead of using `Gettext.PO.Translation` for both regular and pluralized
translation, I introduced the `Gettext.PO.PluralTranslation` struct and changed
a bunch of stuff accordingly. The most important thing that changed is that the
yecc parser now returns a list of `{:translation, translation}` or
`{:plural_translation, translation}` tuples instead of direct translations.