Refactor a lot of interpolation-related code
I created the function `Gettext.Interpolation.to_interpolatable/1`, which
converts a string with "%{interpolations}" in it to a list where those
interpolations are variables. This allowed to simplify the code a lot, since
occurrences of `%{}` patterns are only scanned in one place; if only the
interpolation keys are needed, the list returned by `to_interpolatable/1` can be
filtered rejecting all strings and only keeping atoms.