Always convert interpolation bindings to a map
In the `Gettext` module, interpolation bindings are always used as a map but the
`lgettext/4` and `lngettext/6` functions accept a keyword list as well (for the
syntactic sugar nicety). Instead of checking for the bindings being a list
inside every generated clause, I added a clause for both this functions at the
top of the file with a `when is_list(bindings)` guard, so that this check and
the corresponding conversion to a map happen only in one place.