Make `ln?gettext` accept only maps for interpolation
With this commit, the `ln?gettext` functions (that are injected when a module
uses `Gettext`) stop accepting a keyword list for interpolation, allowing only a
map to be passed. The code that performs the conversion from a keyword list to a
map if the bindings are passed as a keyword list has been moved up to
`Gettext.dgettext/4` and `Gettext.dngettext/6`. This has been possible for two
reasons:

* `ln?gettext` are meant to be private, so we can assume they'll always receive
  maps;
* all the exported APIs pass through `Gettext.d?n?gettext` at some point
  (including the `d?n?gettext` macros injected into modules that use `Gettext`)

Tests have been modified appropriately.