Parse comments without typing them
In this (fat) commit, I changed how the tokenizing and parsing of comments
works. Before this, the tokenizer was able to determine the type of a comment
(reference, translator) and the comments associated with a translation were
stored as a keyword list in the `:comments` field of that translation (e.g.,
`[reference: "foo.ex:1", translator: "Hey there"]`). Now, the tokenizer
tokenizes all comments as the same type of token, and only reference comments
are extracted (by `Gettext.PO.Parser`, after the yecc parsing) into the
`:references` field of the translations.

We'll see how it goes with this solution.