Fix Gettext.PO.Translations.same?/1 with split strings
This function checked for exact equality between the `msgids` (and
`msgid_plurals`) of the translations passed to it; this behaviour is incorrect
since `msgid` and `msgid_plural` can actually be lists of strings representing
strings to concatenate.

We want to compare the concatenated strings, not the lists (we care about the
actual string being equal, not how it's split); this commit does exactly that.