Represent all PO-file strings as lists of binaries
Now all the strings found in PO files are represented as lists of
binaries, regardless of how many there are. Before this commit, if there
was just one string like in this snippet of PO file:

    msgid "foo"

then the string was represented directly as a binary; now it's
represented as a one-element list. This makes things more consistent.