Add the :po_source field to (Plural)Translation
This field is a `{file, line}` tuple which points to where in a PO file the
translation comes from. For now, this will be useful in order to give meaningful
errors when there are duplicated translations but in the future it could be
useful in other ways too.

Note: I had to change a lot of tests from `==` assertions to `=` (match)
assertions because the default value of the `:po_source` field is `nil`, and not
changing these tests would have meant adding the correct `:po_source` field to
all tested translations.