Refactor Gettext.PO.Parser.check_for_duplicates/1
This function now uses throw/catch in order to traverse a list of
translations only once to find duplicates. The code is very contained
and straightforward, so throw/catch shouldn't be a problem here.

The upside of doing this is speed and simplicity; the downside, however,
is that we only find the first duplicate of a translation (instead of
all the duplicates of that translation). For now, this is fine since
duplicate translations should be a not very common case.