Extract the :po_source field out of translations
The `:po_source` field of `(Plural)Translation`s was a `{file, line}`
tuple telling where the translation could be found in the PO file it
came from (if any). The `file` field, however, is much more suited to a
`PO` struct rather than to a translation, so I removed the `:po_source`
field in favour of a `:po_source_line` field for translations and a
`:source_file` field for `PO` structs.