Dump reference comments separately
Before this commit, reference comments in the `:comments` field of a translation
were dumped as regular comments (i.e., with no modification). This behaviour is
incorrect for reference comments, since Gettext for Elixir should handle those
automatically: it attaches the reference to a translation when translations are
extracted from source.

This commit changes the behaviour of dumping a translation
(`Gettext.PO.dump_translation/1`): references are dumped explicitly while
reference comments (`#: ...`) are stripped from the `:comments` field of the
translation.