Refactor the translation extraction
`Gettext.PO.Translations.key/1` allows us to use a different strategy to collect
translations in the `ExtractorAgent`: instead of using a list of translations
and traversing it each time we want to add a new translation to check if a
translation with the same is is there, we're now using a `HashDict`. Much
better in terms of performance and of code simplicity as well.

We're also using the `key/1` function mentioned above in order to sort
translations before writing them to the POT file.