Add some tests for the extraction process
I had some trouble setting up tests for the extraction process as Mix
tasks can be a little annoying to test thoroughly. As a matter of fact,
I didn't test the Mix task at all, but tested the functions from
`Gettext.Extractor` that it calls.

I actually only tested `Gettext.Extract.extracting?/0` (trivial) and
`Gettext.Extract.extract/4`. To test the latter, I compiled a module
with calls to `*gettext` functions on the fly (using
`Code.compile_string/2`) after calling
`Extractor.setup_for_extracting/0` so that translations are extracted
during compilation. I think this strategy works well for now as it's
very simple but allows us to test some of the parts of the extraction
process pretty well.

Sadly, I didn't test `Gettext.Extractor.process_results/0` as that
function writes POT files so it's a little troubling to use it cleanly
in a test environment.