Add the :pot_onto_po option to Gettext.PO.merge/3
If this option is set to true, than the merge/3 function assumes that
the old %PO{} struct is a .po file and the new %PO{} struct is a .pot
file. What this means is that the msgid of matching translations will be
taken from the old translation instead of the new one (since it should
be an empty string in the new one as it comes from a .pot file).