Move some merging functions to Gettext.Extractor
This is the first commit in a series of commits aimed at making
merging of PO/POT/in-memory files better.

Specifically, this commit removes Gettext.PO.merge/2 and
Gettext.PO.Translations.merge/2 and moves their functionality to the
Gettext.Extractor module. This is done as merging in the context of the
extraction is very different from merging in the context of
gettext.merge (which will be tackled in the next commits).