Take care of bad cmdline args in gettext.extract
Before this commit, the task was extracting translation even if the
arguments passed on the command line to `gettext.extract` were wrong.
This means translations were extracted but an error was still returned
by Mix (which I think is confusing).

This way, the extraction is run only if the args are correct.