Fail more gracefully in Gettext.*gettext functions
Right now, if you give, e.g., an atom instead of a binary as the msgid
to `Gettext.dgettext/4`, it blows up way down in implementation
details (in a function from the `Regex` module). This commit adds guards
to the `Gettext.*gettext` functions so that they fail with a
`FunctionClauseError`, which may not be the best in clarity but should
still be much better than what we have now.