Handle unknown locales in Gettext.Plural
Right now, this happens:



  1. A user wants to use a locale that Gettext.Plural doesn't cover,

     e.g., nb-no.

  2. They add the nb-no/LC_MESSAGES/*.po translation files, hence...

  3. ...a nb-no clause is generated for the l(n)gettext functions in

     each backend.

  4. This clause is hit when translating nb-no messages, so we call

     Gettext.Plural.plural("nb-no", n) (if Gettext.Plural is the chosen

     pluralizer).

  5. Gettext.Plural doesn't cover "nb-no", thus function clause error.



We decided to add a catch-all clause to the functions in Gettext.Plural

to handle such cases; while doing that, we also decided to stop handling

locales such as `"en" <> _` and start handling them in a more controlled

way. Now, `"en"` and `"en_US"` match but `"enfoo"` doesn't.