Show the file name in syntax errors when extracting
The current `Gettext.Extractor` code manually reads files through
`File.read!/1` and then parses them with
`Gettext.PO.parse_string!/1`. The problem with this is that if there's a
syntax error, then the file name where the error is won't appear in the
error message as the error is raised by `parse_string!/1`.

This commit fixes this by still calling `File.read!/1`, but also calling
`parse_file!/1` after that (so effectively reading the file twice, which
should be fine as we don't care for raw speed here).