Don't print tokens as Erlang terms in syntax error
yecc prints tokens as Erlang terms (e.g., binaries are `<<"like
this">>`) instead of Elixir terms. This commit introduces a pretty ugly
hack that parses and rewrites yecc error reasons to print some terms as
Elixir terms (like binaries) instead of Erlang terms.

This is ugly, but it's necessary given the lack of extensibility in
yecc; in fact, Elixir itself does this as can be seen here:
https://github.com/elixir-lang/elixir/blob/b806511b887776c1f9ab349a377350b8c9c069f1/lib/elixir/src/elixir_errors.erl#L51-L103.