Add a :compiler_po_wildcard option to the Gettext configuration (#100)
Right now, the :gettext compiler uses a hardcoded default wildcard
of "**/*.po" to find all the PO files under "priv" (so that manifests
can be built and updated). This turns out to be very slow when there are
a lot of files in "priv", as it may happen for example in a Phoenix
application with a lot of stuff in "priv/vendor" (see
https://github.com/phoenixframework/phoenix/issues/1671).

This commit adds support for the :compiler_po_wildcard option to the
Gettext configuration. This option specifies the wildcard that the
compiler uses  to find PO files (relative to "priv"). It defaults to
"gettext/*/LC_MESSAGES/*.po", which is in line with the default
directory for Gettext backends, which is "priv/gettext".