Check for duplicated translations regardless of how msgid is split
Before this, when parsing PO(T) files we were checking for duplicated
translations (with the same msgid or msgid+msgid_plural), but we were comparing
the :msgid/:msgid_plural fields of the translation structs directly. However,
these fields are lists of strings in order to keep the exact same formatting as
the original files when parsing/dumping, so effectively identical msgids could
be considered different because split in different places. For example:

    msgid ""
      "foo"

would be considered different from:

    msgid "foo"
      ""

This commit fixes this by using Gettext.PO.Translations.key/1 to find the id of
the translations.