Simplify Gettext.Interpolation.interpolate/2
This version doesn't take a "tag" that it returns as {tag, interpolated_string}
when interpolation is successful. Instead, it returns either {:ok,
interpolated_string} or {:missing_bindings, incomplete_string, bindings} if
bindings are missing.

This simplifies interpolate/2 both in its implementation as well as its
interface, that is, it now either succeeds or fails, the responsibility of
handling each of those cases is the caller's.