Fix a bug when dumping references (#213)
Closes #212.

If the first reference was longer than the whole line length, we
would write an empty "#:" reference comment because of the way that
dumping of reference comments was built. I changed the implementation
so that we chunk references so that each chunk represents a line to
dump, which fixes the problem described above.