Add configuration option `allowed_locales`
As projects grow, the number of locales and content of each one tend to
grow. We need all the locales in production and maybe on CI, but it
might make the development experience very painful when recompiling all
the locales. This commit adds support to only allow certain locales
configurable by environment