Change references from "country" to "territory" "country" refers to a political entity which can cause conflict when territories are disputed.
A list of disputed territories is available at
https://en.wikipedia.org/wiki/List_of_territorial_disputes

This commit uses the term "territory" rather than
"country" to reflect that locales serve a purpose
to reflect user preferences, not to express a political
agreements.

Annex A to ISO/IEC 15897:1998(E) also refers to "territory"
for submissions.

ISO3166 which lists "country codes" says for "TW" (Taiwan)
that ISO recognises Taiwan as "Taiwan, Province of China".
This view is not widely accepted by the people who live
in Taiwan showing how challenging it is to reflect
both cultural preferences and manage political realities.

Using the term "territory" tends to nullify these issues
and focus on the reason that a locale definition exists.

CLDR (https://cldr.unicode.org) uses the terms "territory"
or "region" interchangeably for this reason.