Keep custom flags with the `custom_flags_to_keep` option
Background:
While merging translations with `mix gettext.merge`
flags get cleared out and replaced with the exception of
`fuzzy`, which is kept.

This change will allow the `custom_flags_to_keep` flag to
add a list of flags to be kept in addition to `fuzzy` which
helps prevent information loss if external tools adds their
own flags to messages.

Fixes #369